Changed the Age property from a double to a decimal

+++ UnitTestProject1/UnitTest1.cs
@@ -23,17 +23,17 @@ namespace UnitTestProject1
         public void TestAdult()
         {
             // Under 18 should return false
-            _p.Age = 17.99;
+            _p.Age = (decimal)17.99;
             bool result = _p.isAdult();
             Assert.IsFalse(result);
 
             // Equal to 18 should return true
-            _p.Age = 18.00;
+            _p.Age = (decimal)18.00;
             result = _p.isAdult();
             Assert.IsTrue(result);
 
             // Over 18 should return true
-            _p.Age = 18.01;
+            _p.Age = (decimal)18.01;
             result = _p.isAdult();
             Assert.IsTrue(result);
         }
